+++ guillotina_cms/content/__init__.py
@@ -1,6 +1,7 @@
 from .document import Document  # noqa
 from .link import Link  # noqa
 from .news import News  # noqa
+from .image import Image  # noqa
 
 from .id_generator import IDGenerator  # noqa
 from . import contraintypes  # noqa