+++ guillotina_cms/interfaces/content.py
@@ -1,5 +1,7 @@
 from guillotina.interfaces import IFolder
+from guillotina.interfaces import IItem
 from guillotina_cms.fields.richtext import RichTextField
+from guillotina.fields import CloudFileField
 
 
 class IDocument(IFolder):
@@ -8,3 +10,10 @@ class IDocument(IFolder):
         title='Text',
         required=False,
         widget='richtext')
+
+class IImage(IItem):
+
+    image = CloudFileField(
+        title='Image',
+        required=False,
+        widget='image')